refactor(quantization): keep Triton fallback in dispatch
