+++ tests/unit_tests/dispatch/test_flaggems_quantization.py
@@ -60,29 +60,15 @@ def test_dynamic_per_token_quant_propagates_failure_for_dispatch_fallback(
         dynamic_per_token_quant_int8_flaggems_vllm(torch.ones((1, 4)))
 
 
-def test_dynamic_per_token_quant_delegates_to_local_triton_fallback(monkeypatch):
-    x = torch.ones((2, 4), dtype=torch.float32)
-    expected_q = torch.ones_like(x, dtype=torch.int8)
-    expected_scale = torch.full((2, 1), 0.25, dtype=torch.float32)
-    calls = []
-
-    def quant_fn(value):
-        calls.append(value)
-        return expected_q, expected_scale
-
-    kernel_module = ModuleType("vllm_fl.quantization.w8a8.flaggems_kernels")
-    kernel_module.dynamic_per_token_quant_int8 = quant_fn
-    monkeypatch.setitem(
-        sys.modules,
-        "vllm_fl.quantization.w8a8.flaggems_kernels",
-        kernel_module,
-    )
-
-    actual_q, actual_scale = dynamic_per_token_quant_int8_flaggems_triton(x)
-
-    assert calls == [x]
-    assert actual_q is expected_q
-    assert actual_scale is expected_scale
+def test_local_triton_quant_validates_input_contract():
+    with pytest.raises(ValueError, match="2D"):
+        dynamic_per_token_quant_int8_flaggems_triton(torch.ones((1, 2, 4)))
+    with pytest.raises(TypeError, match="floating point"):
+        dynamic_per_token_quant_int8_flaggems_triton(
+            torch.ones((1, 4), dtype=torch.int8)
+        )
+    with pytest.raises(ValueError, match="hidden_size"):
+        dynamic_per_token_quant_int8_flaggems_triton(torch.ones((1, 0)))
 
 
 def test_flaggems_quantization_registers_ordered_fallbacks(monkeypatch):
@@ -119,9 +105,6 @@ def test_local_triton_quant_matches_reference_on_cuda(device):
     if device.type != "cuda":
         pytest.skip("local Triton quantization contract is validated on CUDA")
 
-    from vllm_fl.quantization.w8a8.flaggems_kernels import (
-        dynamic_per_token_quant_int8 as triton_quant,
-    )
     from vllm_fl.quantization.w8a8.reference import (
         dynamic_per_token_quant_int8 as reference_quant,
     )
@@ -136,7 +119,7 @@ def test_local_triton_quant_matches_reference_on_cuda(device):
         dtype=torch.float32,
     )
 
-    actual_q, actual_scale = triton_quant(x)
+    actual_q, actual_scale = dynamic_per_token_quant_int8_flaggems_triton(x)
     expected_q, expected_scale = reference_quant(x)
 
     assert torch.equal(actual_q, expected_q)
